refactor: HttpRequestMapper와 HttpRequestParser로 클래스 분리
